requirements.txt
Flask==2.3.4
flask-cors==3.0.10
pyttsx3==2.90
google-generativeai==0.0.6
python-dotenv==1.0.0
